Fixed liquibase for correct work

+++ src/main/java/com/example/basicbookstoreprojectnew/dto/BookDto.java
@@ -1,12 +1,14 @@
 package com.example.basicbookstoreprojectnew.dto;
 
+import java.math.BigDecimal;
+
 public record BookDto(
         Long id,
         String title,
         String author,
         String description,
         String isbn,
-        Integer price,
+        BigDecimal price,
         String coverImage
 ) {
 }
